Add memory storer, add tests.
Add a Storer implementation for in-memory uploads that aren't persisted
anywhere.

Add tests for Upload and Download, and calling the Storer Delete
methods.

Fix our reference to the filetype package to not use gopkg.in anymore,
as Go modules takes care of it and it seems the author has adapted that
stance as well.